Rename packages for consistency
