+++ controllers/reaper/reaper_controller_test.go
@@ -1,4 +1,4 @@
-package controllers
+package reaper
 
 import (
 	"context"
