+++ main.go
@@ -19,13 +19,13 @@ package main
 import (
 	"flag"
 	"fmt"
-	"github.com/k8ssandra/k8ssandra-operator/pkg/reaper"
 	"os"
 
 	cassdcapi "github.com/k8ssandra/cass-operator/apis/cassandra/v1beta1"
 	"github.com/k8ssandra/k8ssandra-operator/pkg/cassandra"
 	"github.com/k8ssandra/k8ssandra-operator/pkg/clientcache"
 	"github.com/k8ssandra/k8ssandra-operator/pkg/config"
+	"github.com/k8ssandra/k8ssandra-operator/pkg/reaper"
 
 	// Import all Kubernetes client auth plugins (e.g. Azure, GCP, OIDC, etc.)
 	// to ensure that exec-entrypoint and run can make use of them.
@@ -42,11 +42,11 @@ import (
 
 	configapi "github.com/k8ssandra/k8ssandra-operator/apis/config/v1beta1"
 	k8ssandraiov1alpha1 "github.com/k8ssandra/k8ssandra-operator/apis/k8ssandra/v1alpha1"
-	reaperv1alpha1 "github.com/k8ssandra/k8ssandra-operator/apis/reaper/v1alpha1"
+	reaperapi "github.com/k8ssandra/k8ssandra-operator/apis/reaper/v1alpha1"
 	replicationapi "github.com/k8ssandra/k8ssandra-operator/apis/replication/v1alpha1"
 	stargateapi "github.com/k8ssandra/k8ssandra-operator/apis/stargate/v1alpha1"
 	k8ssandractrl "github.com/k8ssandra/k8ssandra-operator/controllers/k8ssandra"
-	reapercontrollers "github.com/k8ssandra/k8ssandra-operator/controllers/reaper"
+	reaperctrl "github.com/k8ssandra/k8ssandra-operator/controllers/reaper"
 	replicationctrl "github.com/k8ssandra/k8ssandra-operator/controllers/replication"
 	stargatectrl "github.com/k8ssandra/k8ssandra-operator/controllers/stargate"
 	// +kubebuilder:scaffold:imports
@@ -65,7 +65,7 @@ func init() {
 	utilruntime.Must(replicationapi.AddToScheme(scheme))
 	utilruntime.Must(stargateapi.AddToScheme(scheme))
 	utilruntime.Must(configapi.AddToScheme(scheme))
-	utilruntime.Must(reaperv1alpha1.AddToScheme(scheme))
+	utilruntime.Must(reaperapi.AddToScheme(scheme))
 	// +kubebuilder:scaffold:scheme
 }
 
@@ -197,7 +197,7 @@ func main() {
 		os.Exit(1)
 	}
 
-	if err = (&reapercontrollers.ReaperReconciler{
+	if err = (&reaperctrl.ReaperReconciler{
 		ReconcilerConfig: reconcilerConfig,
 		Client:           mgr.GetClient(),
 		Scheme:           mgr.GetScheme(),
